// Frontend/src/app/model/ticket.ts
export class Ticket {

    transactionId: number | undefined;
    movieName: string | undefined;
    theatreName: string | undefined;
    totalSeats: number | undefined;
    availableSeats: number | undefined;
    bookedSeats: number | undefined;

    // constructor(transactionId: number, movieName: string, theatreName: string, totalSeats: number, availableSeats: number, bookedSeats: number) {
    //     this.transactionId = transactionId;
    //     this.movieName = movieName;
    //     this.theatreName = theatreName;
    //     this.totalSeats = totalSeats;
    //     this.availableSeats = availableSeats;
    //     this.bookedSeats = bookedSeats;
    // }
}