// Frontend/src/app/model/ticket-dto.ts
export class TicketDto {

    transactionId: number | undefined;
    movieName: string | undefined;
    theatreName: string | undefined;
    noOfSeats: number | undefined;

    // constructor(id: number, movieName: string, theatreName: string, noOfSeats: number) {
    //     this.id = id;
    //     this.movieName = movieName;
    //     this.theatreName = theatreName;
    //     this.noOfSeats = noOfSeats;
    // }
}